+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.1.77 May 22, 2020
+
+Added code to combine multiple sequences (see `bin/combine-sequences.py`).
+
 ## 3.1.76 May 3, 2020
 
 Changed `--maskNoCoverage` to `--maskLowCoverage` in `make-consensus.py` and